+++ lib/nfg_ui/components/foundations/image.rb
@@ -5,8 +5,46 @@ module NfgUi
     module Foundations
       # Image doc coming soon
       class Image < NfgUi::Components::Base
+        def css_classes
+          [
+            super,
+            ('img-fluid' if responsive)
+          ].join(' ').squish
+        end
+
+        # Prefer image, image is used
+        # on other components that pull in
+        # the image (e.g. Illustration)
+        def image
+          options.fetch(:image, nil)
+        end
+
         def render
-          content_tag(:img, html_options) {}
+          image_tag(view_context.image_path(image || src || ''), **html_options)
+        end
+
+        # Flag if the image should utilize responsive settings or not.
+        # This typically does thigns like
+        # apply specific css classes (e.g.: 'img-fluid') to the rendered component
+        def responsive
+          options.fetch(:responsive, default_responsive)
+        end
+
+        # Allow :src to come through
+        # since :src is a native and
+        # viable to way source a image path / url
+        def src
+          options.fetch(:src, image)
+        end
+
+        private
+
+        def non_html_attribute_options
+          super.push(:image, :src, :responsive)
+        end
+
+        def default_responsive
+          true
         end
       end
     end